requirements.txt
asgiref==3.8.1
Django==4.1.5
django-webpack-loader==1.8.0
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
sqlparse==0.5.0
typing_extensions==4.11.0
whitenoise==6.6.0
